% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASySD-package.R
\docType{package}
\name{ASySD-package}
\alias{ASySD}
\alias{ASySD-package}
\title{ASySD: Automated Citation Deduplication}
\description{
The Automated Systematic Search Deduplicator ('ASySD') includes functions to automatically detect and remove duplicate citations from literature searches. For details, see our paper "The Automated Systematic Search Deduplicator (ASySD): a rapid, open-source, interoperable tool to remove duplicate citations in biomedical systematic reviews. Hair K / Bahor Z / Macleod M / Liao J / Sena E (2023) \doi{10.1186/s12915-023-01686-z}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://camaradesuk.github.io/ASySD/}
  \item Report bugs at \url{https://github.com/camaradesuk/ASySD/issues}
}

}
\author{
\strong{Maintainer}: Kaitlyn Hair \email{kaitlyn.hair@ed.ac.uk} (\href{https://orcid.org/0000-0003-0180-7343}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Lukas Wallrich \email{lukas.wallrich@gmail.com} (\href{https://orcid.org/0000-0003-2121-5177}{ORCID})
}

}
\keyword{internal}
