% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations_df}
\alias{citations_df}
\title{Example citation data}
\format{
An object of class \code{data.frame} with 1001 rows and 15 columns.
}
\usage{
citations_df
}
\description{
Example citation data
}
\keyword{datasets}
