% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{field_codes_wos}
\alias{field_codes_wos}
\title{Field conversion codes for Web of Science imports}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 4 columns.
}
\usage{
field_codes_wos
}
\description{
Field conversion codes for Web of Science imports
}
\author{
Kaitlyn Hair \email{kaitlyn.hair@ed.ac.uk}
}
\keyword{datasets}
