% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{format_citations}
\alias{format_citations}
\title{This function formats citation data for deduplication}
\usage{
format_citations(raw_citations)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns and id column}
}
\value{
Dataframe of formatted citations with id
}
\description{
This function formats citation data for deduplication
}
