% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{generate_dup_id}
\alias{generate_dup_id}
\title{This function generates a duplicate ID for sets of matching citations}
\usage{
generate_dup_id(true_pairs, formatted_citations)
}
\arguments{
\item{true_pairs}{citation matches which are true duplicates'}

\item{formatted_citations}{formatted citation data}
}
\value{
Dataframe of formatted citation data with duplicate id
}
\description{
This function generates a duplicate ID for sets of matching citations
}
