% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R, R/dedup_refs_new.R
\name{keep_one_unique_citation}
\alias{keep_one_unique_citation}
\title{This function retains one citation in a set of matching records}
\usage{
keep_one_unique_citation(
  raw_citations,
  matched_pairs_with_ids,
  keep_source,
  keep_label
)

keep_one_unique_citation(
  raw_citations,
  matched_pairs_with_ids,
  keep_source,
  keep_label
)
}
\arguments{
\item{raw_citations}{original citation data with ids}

\item{matched_pairs_with_ids}{citation data with duplicate ids}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}
}
\value{
Dataframe of citation data with duplicate citation rows removed

Dataframe of citation data with duplicate citation rows removed
}
\description{
This function retains one citation in a set of matching records

This function retains one citation in a set of matching records
}
