% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_csv_refs}
\alias{load_csv_refs}
\title{Load in csv references file}
\usage{
load_csv_refs(path)
}
\arguments{
\item{path}{Path to the input file}
}
\value{
A dataframe of the Endnote references
}
\description{
This function loads in a comma separated values references file.
}
