% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_refs}
\alias{load_refs}
\title{Load in Endnote references via XML file}
\usage{
load_refs(path, method)
}
\arguments{
\item{path}{Path to the input file}
}
\value{
A dataframe of the Endnote references
}
\description{
This function loads in an Endnote XML file. It assumes that thefile is exported in
standard Endnote XML format.
}
