% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_txt_refs}
\alias{load_txt_refs}
\title{Load in tab delimited reference file}
\usage{
load_txt_refs(path)
}
\arguments{
\item{path}{Path to the input file}
}
\value{
A dataframe of the Endnote references
}
\description{
This function loads in a tab delimited references file. It assumes that thefile is exported in
standard Endnote tab delimited format.
}
