% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{merge_metadata}
\alias{merge_metadata}
\title{This function generates a duplicate ID for sets of matching citations}
\usage{
merge_metadata(matched_data_with_ids, raw_citations_with_id)
}
\arguments{
\item{matched_data_with_ids}{citation data with duplicate ids}

\item{raw_citations_with_id}{original citation data with unique ids}
}
\value{
Dataframe of formatted citation data with duplicate id
}
\description{
This function generates a duplicate ID for sets of matching citations
}
