% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R, R/dedup_refs2.R
\name{order_citations}
\alias{order_citations}
\title{This function orders citation data for deduplication}
\usage{
order_citations(raw_citations)

order_citations(raw_citations)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns and id column}
}
\value{
Dataframe of ordered citations with id

Dataframe of ordered citations with id
}
\description{
This function orders citation data for deduplication

This function orders citation data for deduplication
}
