% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_refs.R
\name{write_citations}
\alias{write_citations}
\title{Write citation data to disk}
\usage{
write_citations(citations, type = c("ris", "txt", "csv", "bib"), filename)
}
\arguments{
\item{citations}{A data frame containing citations - usually post-deduplication}

\item{type}{export type}

\item{filename}{output file name}
}
\value{
No return value, called for side effects (writes file to disk).
}
\description{
This function writes citation data to disk in different formats.
}
\examples{
\dontrun{
write_citations(citations, type = "csv", filename = "output.csv")
}
}
