% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{add_id_citations}
\alias{add_id_citations}
\title{This function adds an id to citation data if missing}
\usage{
add_id_citations(raw_citations)
}
\arguments{
\item{raw_citations}{Citation dataframe with relevant columns}
}
\value{
Dataframe of citations with id
}
\description{
This function adds an id to citation data if missing
}
