% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R, R/dedup_refs_new.R
\name{dedup_citations}
\alias{dedup_citations}
\title{Deduplicate citations}
\usage{
dedup_citations(
  raw_citations,
  manual_dedup = TRUE,
  merge_citations = FALSE,
  keep_source = NULL,
  keep_label = NULL
)

dedup_citations(
  raw_citations,
  manual_dedup = TRUE,
  merge_citations = FALSE,
  keep_source = NULL,
  keep_label = NULL
)
}
\arguments{
\item{raw_citations}{A dataframe containing duplicate ciations}

\item{manual_dedup}{Logical value. Do you want to retrieve dataframe for manual deduplication?}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}
}
\value{
A list of 2 dataframes - unique citations and citations to be manually deduplicated if option selected

A list of 2 dataframes - unique citations and citations to be manually deduplicated if option selected
}
\description{
This function deduplicates citation data

This function deduplicates citation data
}
