% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_citations}
\alias{dedup_citations}
\title{This function deduplicates citation data}
\usage{
dedup_citations(
  raw_citations,
  manual_dedup = TRUE,
  merge_citations = TRUE,
  keep_source = NULL,
  keep_label = NULL,
  extra_merge_fields = NULL,
  shiny_progress = FALSE,
  show_unknown_tags = TRUE,
  user_input = NA
)
}
\arguments{
\item{raw_citations}{A dataframe containing duplicate ciations}

\item{manual_dedup}{Logical value. Do you want to retrieve dataframe for manual deduplication?}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}

\item{extra_merge_fields}{Add additional fields to merge, output will be similar to the label, source, and record_id columns with commas between each merged value}

\item{shiny_progress}{Switch on progress indicators for shiny applications}

\item{show_unknown_tags}{When a label, source, or other merged field is missing, do you want this to show as "unknown"?}
}
\value{
A list of 2 dataframes - unique citations and citations to be manually deduplicated if option selected
}
\description{
This function deduplicates citation data
}
