% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_dedup.R
\name{dedup_citations_add_manual}
\alias{dedup_citations_add_manual}
\title{This function performs additional deduplication with the additional of manually flagged duplicates}
\usage{
dedup_citations_add_manual(
  unique_citations,
  merge_citations = TRUE,
  keep_source = NULL,
  keep_label = NULL,
  additional_pairs,
  extra_merge_fields = NULL,
  show_unknown_tags = TRUE
)
}
\arguments{
\item{unique_citations}{A dataframe containing citations after automated deduplication}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}

\item{additional_pairs}{dataframe of citations with manual pairs, a subset of the manual pairs export. If a \code{result} column is included, only those with a value of \code{match} will be merged}

\item{extra_merge_fields}{Add additional fields to merge, output will be similar to the label, source, and record_id columns with commas between each merged value}

\item{show_unknown_tags}{When a label, source, or other merged field is missing, do you want this to show as "unknown"?}
}
\value{
Unique citations post manual deduplication
}
\description{
This function performs additional deduplication with the additional of manually flagged duplicates
}
\examples{

# Perform deduplication
result <- dedup_citations(citations_df, keep_source="Embase")

# View unique citations
res_unique <- result$unique
head(result$manual_dedup)

true_dups <- result$manual_dedup[1:5,]
# or equivalently
true_dups <- result$manual_dedup
true_dups$result[1:5] <- "match"

# You can also use a Shiny interface to review the potential duplicates
# true_dups <- manual_dedup_shiny(result$manual_dedup)

final_result <- dedup_citations_add_manual(res_unique, additional_pairs = true_dups)

}
