% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_citations_add_manual}
\alias{dedup_citations_add_manual}
\title{Deduplicate citations}
\usage{
dedup_citations_add_manual(
  raw_citations,
  merge_citations = FALSE,
  preferred_source = NULL,
  additional_pairs
)
}
\arguments{
\item{raw_citations}{A dataframe containing duplicate ciations}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{manual_dedup}{Logical value. Do you want to retrieve dataframe for manual deduplication?}

\item{keep_source}{Selected citation label to preferentially retain in the dataset as the unique record}
}
\value{
A list of 2 dataframes - unique citations and citations to be manually deduplicated if option selected
}
\description{
This function deduplicates citation data
}
\examples{

result <- dedup_citations(data, keep_source = "pubmed")
unique_citations <- result$unique
}
