% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R, R/dedup_refs_new.R
\name{dedup_citations_add_manual}
\alias{dedup_citations_add_manual}
\title{Deduplicate citations with manual pairs added}
\usage{
dedup_citations_add_manual(
  raw_citations,
  merge_citations = FALSE,
  keep_source = NULL,
  keep_label = NULL,
  additional_pairs
)

dedup_citations_add_manual(
  raw_citations,
  merge_citations = FALSE,
  keep_source = NULL,
  keep_label = NULL,
  additional_pairs
)
}
\arguments{
\item{raw_citations}{A dataframe containing duplicate citations}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}

\item{additional_pairs}{dataframe of citations with manual pairs, a subset of the manual pairs export}

\item{manual_dedup}{Logical value. Do you want to retrieve dataframe for manual deduplication?}
}
\value{
Unique citations post added manual deduplication

A list of 2 dataframes - unique citations and citations to be manually deduplicated if option selected
}
\description{
This function deduplicates citation data

This function deduplicates citation data
}
