% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_citations_add_manual}
\alias{dedup_citations_add_manual}
\title{This function deduplicates citation data}
\usage{
dedup_citations_add_manual(
  raw_citations,
  merge_citations = FALSE,
  keep_source = NULL,
  keep_label = NULL,
  additional_pairs,
  extra_merge_fields = NULL
)
}
\arguments{
\item{raw_citations}{A dataframe containing duplicate citations}

\item{merge_citations}{Logical value. Do you want to merge matching citations?}

\item{keep_source}{Character vector. Selected citation source to preferentially retain in the dataset as the unique record}

\item{keep_label}{Selected citation label to preferentially retain in the dataset as the unique record}

\item{additional_pairs}{dataframe of citations with manual pairs, a subset of the manual pairs export}

\item{extra_merge_fields}{Add additional fields to merge, output will be similar to the label, source, and record_id columns with commas between each merged value}
}
\value{
Unique citations post added manual deduplication
}
\description{
This function deduplicates citation data
}
