% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{dedup_large_search}
\alias{dedup_large_search}
\title{Get new unique citations and perform deduplication}
\usage{
dedup_large_search(citations, keep_source = NULL, keep_label = NULL)
}
\arguments{
\item{citations}{A dataframe containing citation information.}
}
\value{
A list with components:
\describe{
\item{unique}{Dataframe containing unique citations.}
\item{manual_dedup}{Dataframe containing citations to be manually checked for duplicates.}
}
}
\description{
This function processes a dataframe of citations, deduplicates them, and provides unique citation results.
}
