% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{keep_one_unique_citation}
\alias{keep_one_unique_citation}
\title{This function retains one citation in a set of matching records}
\usage{
keep_one_unique_citation(
  raw_citations_with_id,
  matched_pairs_with_ids,
  preferred_source
)
}
\arguments{
\item{raw_citations_with_id}{original citation data with ids}

\item{matched_pairs_with_ids}{citation data with duplicate ids}
}
\value{
Dataframe of citation data with duplicate citation rows removed
}
\description{
This function retains one citation in a set of matching records
}
