% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_multi_search}
\alias{load_multi_search}
\title{Load in citations for deduplication}
\usage{
load_multi_search(paths, names, method)
}
\arguments{
\item{method}{Import method}

\item{path}{Relative path to the citations file}

\item{name}{File name of input file or files}
}
\value{
A dataframe of the citations
}
\description{
This function loads in a citation file within the shiny app
}
