% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_multi_search}
\alias{load_multi_search}
\title{Load in citations for deduplication}
\usage{
load_multi_search(paths, names, method)
}
\arguments{
\item{paths}{File path to the input file or files}

\item{method}{Loading citations methoddepending on file format}
}
\value{
A dataframe of the Endnote references
}
\description{
This function loads in an Endnote XML file OR csv OR text.
}
