% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_multi_search}
\alias{load_multi_search}
\title{Load in citations for deduplication}
\usage{
load_multi_search(paths, names, method)
}
\arguments{
\item{paths}{Relative paths to the citations file or files}

\item{names}{File names of input file or files}

\item{method}{Import method}
}
\value{
A dataframe of the loaded citations.
}
\description{
This function loads in a citation file within the shiny app
}
\examples{
paths <- c("vignettes/new_sr_search.csv", "vignettes/old_sr_search.csv")
names <- c("new_sr_search.csv", "old_sr_search.csv")
df <- load_multi_search(paths, names, method="csv")
}
