% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_search}
\alias{load_search}
\title{Load in citations for deduplication}
\usage{
load_search(path, method)
}
\arguments{
\item{path}{File path to the input file}

\item{method}{Loading citations methoddepending on file format}

\item{paths}{File path to the input file or files}
}
\value{
A dataframe of the Endnote references

A dataframe of the Endnote references
}
\description{
This function loads in an Endnote XML file OR csv OR text.
}
