% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refs.R
\name{load_search}
\alias{load_search}
\title{Load in citations for deduplication}
\usage{
load_search(path, method)
}
\arguments{
\item{path}{File path to the citations file}

\item{method}{Import method}
}
\value{
A dataframe of the citations
}
\description{
This function loads in an citations file.
}
