% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_dedup_shiny.R
\name{manual_dedup_shiny}
\alias{manual_dedup_shiny}
\title{A Shiny interface to review potential duplicates}
\usage{
manual_dedup_shiny(df, cols = names(df))
}
\arguments{
\item{manual_dedup}{The dataframe with potential duplicates returned from \code{dedup_citations}}
}
\value{
The dataframe with a \code{result} column indicating whether the entry
constitutes a duplicate - to be passed to \code{dedup_citations_add_manual()}
}
\description{
\code{dedup_citations()} can return potential duplicates for manual review. This
function takes these potential duplicates and provides a Shiny interface to
review them and select those that should be deduplicated. The output can then
be passed to \code{dedup_citations_add_manual()} to complete the deduplication, or
be used to call this function again if manual review is not yet complete.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Perform deduplication
result <- dedup_citations(citations_df, keep_source="Embase")

# Manually review potential duplicates
manual_review <- manual_dedup_shiny(result$manual_dedup)

# Complete deduplication
final_result <- dedup_citations_add_manual(result$unique, additional_pairs = manual_review)
\dontshow{\}) # examplesIf}
}
