% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{match_citations}
\alias{match_citations}
\title{This function identifies matching pairs of citations}
\usage{
match_citations(formatted_citations)
}
\arguments{
\item{formatted_citations}{Formatted citation dataframe with relevant columns and id column}
}
\value{
Dataframe of citation pairs
}
\description{
This function identifies matching pairs of citations
}
