% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{merge_metadata}
\alias{merge_metadata}
\title{This function generates a duplicate ID for sets of matching citations}
\usage{
merge_metadata(matched_pairs_with_ids, extra_merge_fields)
}
\arguments{
\item{matched_pairs_with_ids}{citation data with duplicate ids}

\item{extra_merge_fields}{Add additional fields to merge, output will be similar to the label, source, and record_id columns with commas between each merged value}
}
\value{
Dataframe of formatted citation data with duplicate id
}
\description{
This function generates a duplicate ID for sets of matching citations
}
