% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup_refs.R
\name{merge_metadata}
\alias{merge_metadata}
\title{This function generates a duplicate ID for sets of matching citations}
\usage{
merge_metadata(raw_citations, matched_pairs_with_ids, keep_source)
}
\arguments{
\item{raw_citations}{original citation data with unique ids}

\item{matched_pairs_with_ids}{citation data with duplicate ids}
}
\value{
Dataframe of formatted citation data with duplicate id
}
\description{
This function generates a duplicate ID for sets of matching citations
}
