% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_refs.R
\name{write_citations_app}
\alias{write_citations_app}
\title{This function writes citation data to disk in different formats}
\usage{
write_citations_app(citations, type = c("ris", "txt", "csv", "bib"), filename)
}
\arguments{
\item{citations}{A dataframe containing citations - usually post-deduplication}

\item{type}{export type}
}
\value{
file export
}
\description{
This function writes citation data to disk in different formats
}
